% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_Communication_skill.R
\docType{data}
\name{GR_Communication_skill}
\alias{GR_Communication_skill}
\title{Communication skills of L2-Greek learners in the Greek Learner Corpus [GLC]}
\format{A vector with 100 entries of character values.}
\description{
The dataset includes the level of communication skills of 100 students whose language proficiency assessment tests recorded in the GLC.
}
\references{
\href{https://auth.clarin.gr/resources/browse/greek-learner-corpus-uam-corpus-tool-version-3/7d9c64c0c83311e6b526aa3fc6ebde2ceadb41214f994fb9bc568e74a6fe3edb/}{Visit and download} the GLC's site in Clarin:el.
}
\keyword{GLC}
\keyword{datasets}
