% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Package Installation Checking}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{a vector of packages to be checked.}
}
\value{
None
}
\description{
Check to see if packages are installed. Install them if they are not, then load them into the R session.
}
\examples{
packages <- c("ggplot2", "dplyr", "tidyr", "readr", "XML")
ipak(packages)

}
\seealso{
\code{\link{install.packages}}
}
